+++ lib/models/segment.dart
@@ -1,9 +1,6 @@
 part of '../segmented_control_plus.dart';
 
 class Segment<T> {
-  /// Icon of the segment
-  IconData icon;
-
   /// The value passed on onTap
   T value;
 
@@ -11,8 +8,29 @@ class Segment<T> {
   String? tooltip;
 
   Segment({
-    required this.icon,
     required this.value,
     this.tooltip,
+  });
+}
+
+class SegmentIcon<T> extends Segment<T> {
+  /// Icon of the segment
+  IconData icon;
+
+  SegmentIcon({
+    required this.icon,
+    required super.value,
+    super.tooltip,
+  });
+}
+
+class SegmentIconBuilder<T> extends Segment<T> {
+  /// Icon of the segment
+  Widget Function(BuildContext context, Color color) iconBuilder;
+
+  SegmentIconBuilder({
+    required this.iconBuilder,
+    required super.value,
+    super.tooltip,
   });
 }