+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.1.0
+
+- **Breaking Change:** Segment model is now SegmentIcon or SegmentIconBuilder.
+- Added assert to make sure the segments list is not empty.
+
 ## 0.0.3
 
 - Added icon color detection based on what the selected color or background color is.